+++ app/src/main/res/menu/bottom_nav_menu.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+    <item
+        android:id="@+id/nav_answer"
+        android:icon="@android:drawable/ic_menu_help"
+        android:title="Answer"/>
+    <item
+        android:id="@+id/nav_translate"
+        android:icon="@android:drawable/ic_menu_search"
+        android:title="Translate"/>
+    <item
+        android:id="@+id/nav_docs"
+        android:icon="@android:drawable/ic_menu_gallery"
+        android:title="Docs"/>
+</menu>
